services/dataflow/requirements.txt
unidecode==1.3.8
requests==2.32.3
pandas==2.2.3
apache_beam==2.62.0
google-cloud-secret-manager==2.22.1
google-cloud-storage==3.0.0
tqdm==4.67.1